requirements.txt
apipkg==1.4
autopep8==1.3.2
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
colorama==0.3.9
coverage==4.4.1
decorator==4.1.2
Django==1.11.4
django-filter==1.0.4
djangorestframework==3.6.4
execnet==1.4.1
feedgenerator==1.9
feedparser==5.2.1
Flask==0.12.2
gunicorn==19.7.1
httpbin==0.5.0
httplib2==0.10.3
icalendar==3.11.6
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
lxml==3.8.0
MarkupSafe==1.0
oauth2==1.9.0.post1
pep8==1.7.0
py==1.4.34
pycodestyle==2.3.1
pytest==3.2.1
pytest-cov==2.5.1
pytest-forked==0.2
pytest-xdist==1.19.1
python-dateutil==2.6.1
python-dotenv==0.6.5
pytz==2017.2
requests==2.18.3
six==1.10.0
urllib3==1.22
Werkzeug==0.12.2
whitenoise==3.3.0
